Wrap database update in an explicit transaction
All changes to the database in SQLite must be done within a
transaction, so if there isn't one in progress one will be started
implicitly, i.e. each change will be done in a separate transaction.

Since committing a transaction is a fairly expensive operation - on
some filesystems ridiculously expensive - it should be done as
sparingly as possible.

In tests on my Galaxy S, this change makes the update between 2500%
and 4500% faster (for slightly over 100 applications).