Handle rotation in AppDetails activity
This moves the download thread from an anonymous inner class to the
class Downloader, which isn't tied to the activity.

The state of the progress dialog is updated by polling the download
thread at regular intervals - message passing seemed to need a bunch
of synchronized methods and shared variables anyway (to ensure that
all messages reached the right activity) so it was much simpler to
just take it out. But maybe my approach is completely backward.