Also show "add repo" and "update/install" in action bar.
The "update" icon was a "+" (add icon) which looked a bit weird, so a
quick search of other UI's shows that the "refresh" icon is used for
"update". The official Android reference docs say to copy icons that
aren't part of the android.R.drawable.* constants into your own
drawable-* folders, so I've done that here. The icons are from android
SDK version 17.