Fixed bug I introduced depending on later API.
Also added utility method to make checking a bit easier,
and removed reference to SDK from DB (it mentioned in the
comments that SDK_INT was only available in v5, but the
Android docs say it was introduced in v4. Because FDroid
now depends on the Android support library, which in turn
depends on v4, it sould be okay to depen on this.