Refactored API dependent implementations into classes.
See http://stackoverflow.com/a/6495399.

I thought that I could just wrap
API dependent code in an if statement, ant it would only have a problem
if it tried to execute a particular function at runtime. However when
testing on a 1.6 emulator, I was getting "VerifyErrors" which as the
link above suggest, are because it is verifying every statement in a
class. Refactoring out to another class solves this because it only
verifies classes which are loaded at runtime.