Fix issue 231 - action bar drop down not updating.
Requires a bit of a dodgey hack because of a bug in Android,
but it gets the job done. The good thing is that the hack is
quarantined to a class specifically dealing with the ActionBar
for devices that support it, and not required for earlier
devices. In the future if the bug is fixed, it will be relatively
easy to make sure that it is only applied to the versions which
require it.