Fixed #240 - refresh view after resuming FDroid activity.
The problem was that since the changes to implement tab swiping and
fragments, the process for refreshing the lists was changed. The
previous infrastructure for invalidating the result of
FDroidApp.getApps() is still useful, so we presume that the list of apps
is only going to take a few cycles to return, instead of several
seconds. If it has been invalidated, so be it, we will put up with
waiting the few seconds.