Remove "reset" from the preferences page.
Currently, removing the cache does the job much better. The only thing
clearing the cache doesn't do which "reset" did is removing the databases. But
we don't want to do that anyway.

If someone wants to do that, they can just deactivate or remove a repo, and it
will get wiped from the database.