Safer preference managing. Fix "compact layout requires reload".
Previously, everybody had to remember the preference name and the
default value. If it was ever changed, this would have to be updated
everywhere. Now, the Preferences class is responsible for talking to the
SharedPreferences functionality of ANdroid.

I've started with just the compactlayout preference, because
that is what I required for this fix.