try to find ManageRepo's Fragment before recreating it
This helps with dealing with listFragment being null after a rotation. That
was happening when an Intent was coming in from a QR Code scanner, and
the screen was rotated during the process.  This snippet was taken from
SearchResults.java

This is Peter Serwylo's idea, I just typed it up and tested it :)